fix: rename folder (lower-case)
