+++ src/components/container/Container.stories.tsx
@@ -1,7 +1,7 @@
 import {Meta} from "@storybook/react";
 import Container from "./Container";
 import Card from "../card/Card";
-import Text from "../Text/Text";
+import Text from "../text/Text";
 import Badge from "../badge/Badge";
 import ButtonGroup from "../button-group/ButtonGroup";
 import Button from "../button/Button";
