+++ src/components/flex/Flex.stories.tsx
@@ -1,7 +1,7 @@
 import {Meta} from "@storybook/react";
 import React from "react";
 import Flex from "./Flex";
-import Text from "../Text/Text";
+import Text from "../text/Text";
 
 const meta: Meta = {
     title: "Flex",
