+++ src/components/quote/Quote.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Card, {CardType} from "../card/Card";
 import "./Quote.style.scss"
-import Text from "../Text/Text";
+import Text from "../text/Text";
 import CardSection from "../card/CardSection";
 
 export interface QuoteType extends Omit<CardType, "children"> {
